demasiadas cosas, skills, items, iconos, ...
sigo aca abajo D:
... localizations texts, escena, limitador de movimiento, render de invalid path, mejora del control sobre donde podemos pisar, borrado de cosas inecesarias, añadido el tema de skills buttons, combat ui,, mejorado turnos, es mu tarde y el café ya no me hace efecto

bona nit ;3

+++ Assets/Scripts/Combat/ModPerk.cs
@@ -1,6 +1,6 @@
 
 public abstract class ModPerk : Perk {
 
-    public float _modifier;
+    public int _modifier;
 
 }
